New features for Autonomic + Some bug fixes

+++ core/roboconf-dm/src/main/java/net/roboconf/dm/internal/autonomic/RuleBasedEventHandler.java
@@ -27,9 +27,14 @@ package net.roboconf.dm.internal.autonomic;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.ArrayList;
+import java.util.Date;
+import java.util.HashMap;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Properties;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.logging.Logger;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -45,6 +50,7 @@ import javax.mail.internet.MimeMessage;
 import net.roboconf.core.Constants;
 import net.roboconf.core.model.beans.Component;
 import net.roboconf.core.model.beans.Instance;
+import net.roboconf.core.model.beans.Instance.InstanceStatus;
 import net.roboconf.core.model.helpers.ComponentHelpers;
 import net.roboconf.core.model.helpers.InstanceHelpers;
 import net.roboconf.core.utils.Utils;
@@ -65,12 +71,22 @@ public class RuleBasedEventHandler {
 
 	static final String DELETE_SERVICE = "delete-service";
 	static final String REPLICATE_SERVICE = "replicate-service";
+	static final String REPLICATE_INSTANCE = "replicate-instance";
 	static final String MAIL = "mail";
 
 	static final String AUTONOMIC_MARKER = "autonomic";
 
-	private final Manager manager;
 	private final Logger logger = Logger.getLogger( getClass().getName());
+	private final Manager manager;
+	private final AtomicInteger vmCount = new AtomicInteger( 0 );
+
+	boolean disableChecks = false;
+
+	/**
+	 * This map acts as a cache whose key is bound to an application and a given reaction ID.
+	 */
+	private final Map<String,Instance> reactionKeyToLastInstance;
+	private final Map<String,Long> reactionKeyToLastExecution;
 
 
 	/**
@@ -79,6 +95,8 @@ public class RuleBasedEventHandler {
 	 */
 	public RuleBasedEventHandler( Manager manager ) {
 		this.manager = manager;
+		this.reactionKeyToLastInstance = new HashMap<> ();
+		this.reactionKeyToLastExecution = new HashMap<> ();
 	}
 
 
@@ -93,20 +111,96 @@ public class RuleBasedEventHandler {
 			AutonomicRule rule = rules.get( event.getEventId());
 			this.logger.fine( "Autonomic management event. Event ID = " + event.getEventId());
 
+			// This key is bound to an application name and a reaction ID.
+			String reactionKey = ma.getName() + "_";
+			if( rule != null )
+				reactionKey += rule.getReactionId();
+
+			// Maximum number of machine the autonomic can create
+			int maxVmCount = this.manager.getAutonomicMaxRoots();
+			if( maxVmCount == -1 )
+				maxVmCount = Integer.MAX_VALUE;
+
+			// Permissions checks are handled entirely in this method.
+			// Dealing with it in the (several) specific methods would make unit tests more complex.
 			if( rule == null )
 				this.logger.fine( "No rule was found to handle events with the '" + event.getEventId() + "' ID." );
 
 			// EVENT_ID ReplicateService ComponentTemplate
-			else if( REPLICATE_SERVICE.equalsIgnoreCase( rule.getReactionId()))
-				createInstances( ma, rule.getReactionInfo());
+			else if( REPLICATE_SERVICE.equalsIgnoreCase( rule.getReactionId())) {
+
+				this.logger.fine( "Autonomic management: about to create a new instance based on '" + rule.getReactionInfo() + "'." );
+				if( this.vmCount.get() >= maxVmCount ) {
+					this.logger.info( "Autonomic management: the maximum number of instances created by the autonomic has been reached. Service replication is cancelled." );
+
+				} else if( ! checkTimingForAdditions( reactionKey, rule.getDelay())) {
+					this.logger.info( "Autonomic management: the " + rule.getDelay() + "s period has not yet completed since the last machine was created by the autonomic." );
+
+				} else if( checkPermission( reactionKey, true )) {
+					Instance inst = createInstances( ma, rule.getReactionInfo());
+					ack( inst, reactionKey );
+
+				} else {
+					this.logger.warning(
+							"Autonomic management: replication of service of '" + rule.getReactionId() + "' is cancelled. "
+							+ "A previous execution is still in progress (reaction ID = " + rule.getReactionId() + ")." );
+				}
+			}
+
+			// EVENT_ID ReplicateInstance RootInstanceName
+			else if( REPLICATE_INSTANCE.equalsIgnoreCase( rule.getReactionId())) {
+
+				this.logger.fine( "Autonomic management: about to replicate instance '/" + rule.getReactionInfo() + "'." );
+				if( this.vmCount.get() >= maxVmCount ) {
+					this.logger.info( "Autonomic management: the maximum number of instances created by the autonomic has been reached. Instance replication is cancelled." );
+
+				} else if( ! checkTimingForAdditions( reactionKey, rule.getDelay())) {
+					this.logger.info( "Autonomic management: the " + rule.getDelay() + "s period has not yet completed since the last machine was created by the autonomic." );
+
+				} else if( checkPermission( reactionKey, true )) {
+					Instance inst = replicateInstance( ma, rule.getReactionInfo());
+					ack( inst, reactionKey );
+
+				} else {
+					this.logger.warning(
+							"Autonomic management: replication of root instance '" + rule.getReactionId() + "' is cancelled. "
+							+ "A previous execution is still in progress (reaction ID = " + rule.getReactionId() + ")." );
+				}
+			}
 
 			// EVENT_ID StopService ComponentName
-			else if( DELETE_SERVICE.equalsIgnoreCase( rule.getReactionId()))
-				deleteInstances( ma, rule.getReactionInfo());
+			else if( DELETE_SERVICE.equalsIgnoreCase( rule.getReactionId())) {
+
+				this.logger.fine( "Autonomic management: about to delete an instance of '" + rule.getReactionInfo() + "'." );
+				if( ! checkTimingForOthers( reactionKey, rule.getDelay())) {
+					this.logger.info( "Autonomic management: the " + rule.getDelay() + "s period has not yet completed since the last machine was deleted by the autonomic." );
+
+				} else if( checkPermission( reactionKey, false )) {
+					Instance inst = deleteInstances( ma, rule.getReactionInfo());
+					ack( inst, reactionKey );
+
+				} else {
+					this.logger.warning(
+							"Autonomic management: deletion of an instance of '" + rule.getReactionId() + "' is cancelled. "
+							+ "A previous execution is still in progress (reaction ID = " + rule.getReactionId() + ")." );
+				}
+			}
 
 			// EVENT_ID Mail DestinationEmail
-			else if( MAIL.equalsIgnoreCase( rule.getReactionId()))
-				sendEmail(ma, rule.getReactionInfo());
+			else if( MAIL.equalsIgnoreCase( rule.getReactionId())) {
+
+				// FIXME: as we do not really test e-mail notifications, we should find a better
+				// code organization so that we can at least delays on this kind of rule.
+				// As an example, we could extract the e-mail sender in another class we could mock.
+				this.logger.fine( "Autonomic management: about to send an e-mail." );
+				if( ! checkTimingForOthers( reactionKey, rule.getDelay())) {
+					this.logger.info( "Autonomic management: the " + rule.getDelay() + "s period has not yet completed since the last e-mail was sent by the autonomic." );
+
+				} else {
+					sendEmail(ma, rule.getReactionInfo());
+					this.reactionKeyToLastExecution.put( reactionKey, new Date().getTime());
+				}
+			}
 
 			// EVENT_ID Log LogMessage
 			// And default behavior...
@@ -128,8 +222,6 @@ public class RuleBasedEventHandler {
 	 * @throws IOException if the mail properties could not be read
 	 */
 	void sendEmail( ManagedApplication ma, String emailData ) throws IOException {
-
-		this.logger.fine( "Autonomic management: about to send an e-mail." );
 
 		/*
 		 * Sample properties:
@@ -186,7 +278,7 @@ public class RuleBasedEventHandler {
 			message.setSubject(subject);
 			message.setText(data.trim());
 
-			// Send email !
+			// Send email!
 			Transport.send(message);
 
 		} catch( MessagingException e ) {
@@ -203,10 +295,12 @@ public class RuleBasedEventHandler {
 	 * <p>
 	 * (e.g. VM_name/Software_container_name/App_artifact_name)
 	 * </p>
+	 *
+	 * @return the created root instance (can be null in case of error)
 	 */
-	void createInstances( ManagedApplication ma, String componentTemplates ) {
+	Instance createInstances( ManagedApplication ma, String componentTemplates ) {
 
-		this.logger.fine( "Autonomic management: about to create a new instance based on '" + componentTemplates + "'." );
+		Instance result = null;
 		try {
 			if( componentTemplates.startsWith( "/" ))
 				componentTemplates = componentTemplates.substring( 1 );
@@ -228,9 +322,10 @@ public class RuleBasedEventHandler {
 				// compToInstantiate should never be null (check done above).
 
 				// All the root instances must have a different name. Others do not matter.
+				// Use nanoTime() for generated names, as milliseconds can result in test failures (name conflicts).
 				String instanceName;
 				if( previousInstance == null )
-					instanceName = compToInstantiate.getName() + "_" + System.currentTimeMillis();
+					instanceName = compToInstantiate.getName() + "_" + System.nanoTime();
 				else
 					instanceName = compToInstantiate.getName().toLowerCase();
 
@@ -244,10 +339,57 @@ public class RuleBasedEventHandler {
 			rootInstance.data.put( AUTONOMIC_MARKER, "true" );
 			this.manager.instancesMngr().deployAndStartAll( ma, rootInstance );
 
+			// Only update the result if everything went fine
+			result = rootInstance;
+			this.vmCount.incrementAndGet();
+
 		} catch( Exception e ) {
 			this.logger.warning( "The creation of instances (autonomic context) failed. " + e.getMessage());
 			Utils.logException( this.logger, e );
 		}
+
+		return result;
+	}
+
+
+	/**
+	 * Instantiates a new VM from another "template" instance.
+	 * @param ma the managed application
+	 * @param rootInstanceName the name of the root "template" instance
+	 * @return the created root instance (can be null in case of error)
+	 */
+	Instance replicateInstance( ManagedApplication ma, String rootInstanceName ) {
+
+		Instance result = null;
+		try {
+			// Find the instance to copy
+			Instance rootInstanceToCopy = InstanceHelpers.findInstanceByPath( ma.getApplication(), "/" + rootInstanceName );
+			if( rootInstanceToCopy == null )
+				throw new IOException( "Instance " + rootInstanceName + " was not found in application " + ma.getApplication().getName());
+
+			// Copy it
+			Instance copy = InstanceHelpers.replicateInstance( rootInstanceToCopy );
+
+			// Use nanoTime() for generated names, as milliseconds can result in test failures (name conflicts).
+			copy.setName( copy.getComponent().getName() + "_" + System.nanoTime());
+
+			// Register it in the model
+			this.manager.instancesMngr().addInstance( ma, null, copy );
+
+			// Now, deploy and start all
+			copy.data.put( AUTONOMIC_MARKER, "true" );
+			this.manager.instancesMngr().deployAndStartAll( ma, copy );
+
+			// Only update the result if everything went fine
+			result = copy;
+			this.vmCount.incrementAndGet();
+
+		} catch( Exception e ) {
+			this.logger.warning( "The replication of an instance (autonomic context) failed. " + e.getMessage());
+			Utils.logException( this.logger, e );
+		}
+
+		return result;
 	}
 
 
@@ -259,10 +401,11 @@ public class RuleBasedEventHandler {
 	 *
 	 * @param ma the managed application
 	 * @param componentName the component name of an instance to delete
+	 * @return the removed root instance (can be null in case of error)
 	 */
-	void deleteInstances( ManagedApplication ma, String componentName ) {
+	Instance deleteInstances( ManagedApplication ma, String componentName ) {
 
-		this.logger.fine( "Autonomic management: about to delete an instance of '" + componentName + "'." );
+		Instance result = null;
 		try {
 			// Find an instance which was created by the autonomic.
 			// Its root instance is annotated with the "autonomic" marker.
@@ -283,13 +426,162 @@ public class RuleBasedEventHandler {
 			if( instanceToRemove != null ) {
 				this.manager.instancesMngr().undeployAll( ma, instanceToRemove );
 				this.manager.instancesMngr().removeInstance( ma, instanceToRemove );
+
+				// Only update the result if everything went fine
+				result = instanceToRemove;
+				this.vmCount.decrementAndGet();
 			}
 
 		} catch( Exception e ) {
 			this.logger.warning( "The deletion of an instance (autonomic context) failed. " + e.getMessage());
 			Utils.logException( this.logger, e );
 		}
+
+		return result;
 	}
+
+
+	/**
+	 * Checks some actions are allowed.
+	 * <p>
+	 * This method should only invoked for creation and deletion of instances.
+	 * </p>
+	 *
+	 * @param reactionKey the reaction key
+	 * @param add true if an instance is about to be added and started, false if it is about to be removed
+	 * @return true if the reaction can be triggered, false otherwise
+	 */
+	boolean checkPermission( String reactionKey, boolean add ) {
+
+		boolean permitted = true;
+		Instance lastInstance = this.reactionKeyToLastInstance.get( reactionKey );
+
+		// Disabling checks is used in some tests
+		if( ! this.disableChecks && lastInstance != null ) {
+
+			for( Iterator<Instance> it = InstanceHelpers.buildHierarchicalList( lastInstance ).iterator(); it.hasNext() && permitted; ) {
+				Instance inst = it.next();
+
+				// If we want to add a new instance, then all the instances "before"
+				// must have been started.
+				if( add && inst.getStatus() != InstanceStatus.DEPLOYED_STARTED ) {
+					permitted = false;
+					this.logger.warning( "Autonomic management: instance " + inst + " is not yet started (context = " + reactionKey + ")." );
+				}
+
+				// Otherwise, if we want to remove an instance, then all the instances "before"
+				// must have been undeployed.
+				else if( ! add && inst.getStatus() != InstanceStatus.NOT_DEPLOYED ) {
+					permitted = false;
+					this.logger.warning( "Autonomic management: instance " + inst + " is not yet undeployed (context = " + reactionKey + ")." );
+				}
+			}
+
+			// Clean the cache
+			if( permitted )
+				this.reactionKeyToLastInstance.remove( reactionKey );
+		}
+
+		return permitted;
+	}
+
+
+	/**
+	 * Checks time-related data to verify an action can be executed.
+	 * <p>
+	 * These verifications are only meaningful for reactions that create new instances.
+	 * </p>
+	 * <p>
+	 * When no timing information is available, consider it is permitted.
+	 * </p>
+	 *
+	 * @param reactionKey the reaction key
+	 * @param delay the delay to respect (in seconds)
+	 * @return true if the action is permitted, false otherwise
+	 */
+	boolean checkTimingForAdditions( String reactionKey, long delay ) {
+
+		boolean permitted = true;
+		Instance lastInstance = this.reactionKeyToLastInstance.get( reactionKey );
+
+		// Disabling checks is used in some tests
+		if( ! this.disableChecks && lastInstance != null ) {
+
+			String timeAS = lastInstance.data.get( Instance.RUNNING_FROM );
+			if( timeAS != null ) {
+				long time = new Date().getTime() - Long.parseLong( timeAS );
+
+				// >= because delay can be 0
+				permitted = time >= delay * 1000;
+			}
+		}
+
+		return permitted;
+	}
+
+
+	/**
+	 * Checks time-related data to verify an action can be executed.
+	 * <p>
+	 * These verifications are only meaningful for reactions that delete
+	 * instances or send notifications.
+	 * </p>
+	 * <p>
+	 * When no timing information is available, consider it is permitted.
+	 * </p>
+	 *
+	 * @param reactionKey the reaction key
+	 * @param delay the delay to respect (in seconds)
+	 * @return true if the action is permitted, false otherwise
+	 */
+	boolean checkTimingForOthers( String reactionKey, long delay ) {
+
+		boolean permitted = true;
+		if( ! this.disableChecks ) {
+			Long time = this.reactionKeyToLastExecution.get( reactionKey );
+			if( time == null )
+				time = 0L;
+
+			// >= because delay can be 0
+			permitted = (new Date().getTime() - time) >= (delay * 1000);
+		}
+
+		return permitted;
+	}
+
+
+	/**
+	 * Acknowledges an instance was created or removed for a given reaction ID.
+	 * @param instance the resulting instance (can be null)
+	 * @param reactionKey a reaction key (not null)
+	 */
+	void ack( Instance instance, String reactionKey ) {
+
+		if( instance == null )
+			return;
+
+		// Update the cache
+		this.reactionKeyToLastInstance.put( reactionKey, instance );
+
+		// Update the last execution time
+		this.reactionKeyToLastExecution.put( reactionKey, new Date().getTime());
+
+		// Clean it too
+		// Indeed, some actions may invalidate it. As an example, it is not worth
+		// keeping an instance cached for a replicate reaction if it was removed.
+
+		// We here consider an instance can only be cached for only one reaction ID.
+		List<String> reactionKeys = new ArrayList<> ();
+		for( Map.Entry<String,Instance> entry : this.reactionKeyToLastInstance.entrySet()) {
+			if( entry.getValue().equals( instance ))
+				reactionKeys.add( entry.getKey());
+		}
+
+		reactionKeys.remove( reactionKey );
+		for( String key : reactionKeys )
+			this.reactionKeyToLastInstance.remove( key );
+	}
+
 
 	/**
 	 * @author Pierre-Yves Gibello - Linagora
